+++ python/mirascope/llm/tools/__init__.py
@@ -4,11 +4,10 @@ from .async_context_tool import AsyncContextTool
 from .async_tool import AsyncTool
 from .base_tool import BaseTool
 from .context_tool import ContextTool
-from .context_toolkit import ContextToolkit
 from .decorator import ContextToolDecorator, ToolDecorator, tool
 from .tool import Tool
 from .tool_typevars import AgentToolT, ContextToolT, ToolT
-from .toolkit import Toolkit
+from .toolkit import ContextToolkit, Toolkit, ToolkitT
 
 __all__ = [
     "AgentToolT",
@@ -23,5 +22,6 @@ __all__ = [
     "ToolDecorator",
     "ToolT",
     "Toolkit",
+    "ToolkitT",
     "tool",
 ]
